CLUSTER02TEAM08
Tourism helps to “enhance employment opportunities and earnings, which can be of major economic significance to the local population” .It includes many businesses that fall under this large umbrella, such as hotels, motels, resorts, restaurants, theme parks, and much more. The hospitality industry is a broad group of businesses that provide services to customers. The industry can be broken down into three basic areas: accommodations, food and beverage, and travel and tourism.